model/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pypi_0
apted=1.0.3=pypi_0
argon2-cffi=20.1.0=pypi_0
astunparse=1.6.3=pypi_0
async-generator=1.10=pypi_0
attrs=21.2.0=pypi_0
backcall=0.2.0=pypi_0
bleach=3.3.1=pypi_0
ca-certificates=2021.4.13=h06a4308_1
cached-property=1.5.2=pypi_0
cachetools=4.2.1=pypi_0
certifi=2020.12.5=py36h06a4308_0
cffi=1.14.6=pypi_0
chardet=4.0.0=pypi_0
colorama=0.4.4=pypi_0
contextlib2=0.6.0.post1=pypi_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pypi_0
easydict=1.9=pypi_0
editdistance=0.6.0=pypi_0
efficientnet-pytorch=0.7.1=pypi_0
entrypoints=0.3=pypi_0
flatbuffers=1.12=pypi_0
gast=0.3.3=pypi_0
google-auth=1.28.1=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.32.0=pypi_0
h5py=2.10.0=pypi_0
hdf5storage=0.1.18=pypi_0
idna=2.10=pypi_0
imageio=2.9.0=pypi_0
importlib-metadata=3.10.1=pypi_0
ipykernel=5.5.5=pypi_0
ipython=7.16.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.3=pypi_0
jedi=0.18.0=pypi_0
jieba=0.42.1=pypi_0
jinja2=3.0.1=pypi_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=6.1.12=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter-core=4.7.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
lmdb=0.98=pypi_0
lxml=4.6.3=pypi_0
markdown=3.3.4=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.3.4=pypi_0
mistune=0.8.4=pypi_0
ml-collections=0.1.0=pypi_0
nbclient=0.5.3=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pypi_0
networkx=2.5.1=pypi_0
nicelogger=2.0=pypi_0
nltk=3.3=pypi_0
notebook=6.4.0=pypi_0
numpy=1.19.5=pypi_0
oauthlib=3.1.0=pypi_0
opencv-python-headless=4.1.2.30=pypi_0
openssl=1.1.1k=h27cfd23_0
opt-einsum=3.3.0=pypi_0
packaging=21.0=pypi_0
pandas=1.1.5=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.8.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.1.2=pypi_0
pip=21.0.1=py36h06a4308_0
prometheus-client=0.11.0=pypi_0
prompt-toolkit=3.0.19=pypi_0
protobuf=3.15.8=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pypi_0
pygments=2.9.0=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.18.0=pypi_0
python=3.6.13=hdb3f193_0
python-dateutil=2.8.1=pypi_0
python-docx=0.8.10=pypi_0
python-levenshtein=0.12.2=pypi_0
pytz=2021.3=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.1.0=pypi_0
qtconsole=5.1.1=pypi_0
qtpy=1.9.0=pypi_0
readline=8.1=h27cfd23_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-image=0.17.2=pypi_0
scipy=1.5.4=pypi_0
seaborn=0.11.2=pypi_0
send2trash=1.7.1=pypi_0
setuptools=52.0.0=py36h06a4308_0
six=1.15.0=pypi_0
sqlite=3.35.4=hdfb4753_0
tensorboard=2.4.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorboardx=2.4=pypi_0
tensorflow=2.4.1=pypi_0
tensorflow-estimator=2.4.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.10.1=pypi_0
testpath=0.5.0=pypi_0
thop=0.0.31-2005241907=pypi_0
tifffile=2020.9.3=pypi_0
tk=8.6.10=hbc83047_0
torch=1.3.1=pypi_0
torchvision=0.4.2=pypi_0
tornado=6.1=pypi_0
tqdm=4.60.0=pypi_0
traitlets=4.3.3=pypi_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.26.4=pypi_0
utils=1.0.1=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
wrapt=1.12.1=pypi_0
xlrd=2.0.1=pypi_0
xlwt=1.3.0=pypi_0
xz=5.2.5=h7b6447c_0
zipp=3.4.1=pypi_0
zlib=1.2.11=h7b6447c_3
